// frontend/src/api/model/githubCom4H1RZooraInternalDomainTicketStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainTicketStatus = typeof GithubCom4H1RZooraInternalDomainTicketStatus[keyof typeof GithubCom4H1RZooraInternalDomainTicketStatus];


export const GithubCom4H1RZooraInternalDomainTicketStatus = {
  TicketStatusOpen: 'open',
  TicketStatusAnswered: 'answered',
  TicketStatusClosed: 'closed',
} as const;