// frontend/src/api/model/githubCom4H1RZooraInternalDomainTicketType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainTicketType = typeof GithubCom4H1RZooraInternalDomainTicketType[keyof typeof GithubCom4H1RZooraInternalDomainTicketType];


export const GithubCom4H1RZooraInternalDomainTicketType = {
  TicketTypeQuestion: 'question',
  TicketTypeGradeObjection: 'grade_objection',
  TicketTypeOther: 'other',
} as const;